+++ src/main/java/gift/controller/admin/AdminController.java
@@ -1,0 +1,17 @@
+package gift.controller.admin;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/admin")
+public class AdminController {
+
+    @GetMapping("")
+    public String adminHome(Model model) {
+        model.addAttribute("message", "관리자 페이지");
+        return "admin";
+    }
+}
