+++ src/main/java/gift/controller/admin/ProductAdminController.java
@@ -1,13 +1,9 @@
-package gift.controller;
+package gift.controller.admin;
 
 import gift.dto.category.CategoryResponse;
-import gift.dto.member.MemberRequest;
-import gift.dto.member.MemberResponse;
 import gift.dto.product.ProductRequest;
 import gift.dto.product.ProductResponse;
-import gift.model.Member;
 import gift.service.CategoryService;
-import gift.service.MemberService;
 import gift.service.ProductService;
 import jakarta.validation.Valid;
 import java.util.List;
@@ -27,28 +23,18 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/admin")
-public class AdminController {
+@RequestMapping("/admin/products")
+public class ProductAdminController {
 
     private final ProductService productService;
-    private final MemberService memberService;
     private final CategoryService categoryService;
 
-    public AdminController(ProductService productService, MemberService memberService,
-        CategoryService categoryService) {
+    public ProductAdminController(ProductService productService, CategoryService categoryService) {
         this.productService = productService;
-        this.memberService = memberService;
         this.categoryService = categoryService;
     }
 
     @GetMapping("")
-    public String adminHome(Model model) {
-        model.addAttribute("message", "관리자 페이지");
-        return "admin";
-    }
-
-    // 상품 관리
-    @GetMapping("/products")
     public String getAllProducts(Model model,
         @PageableDefault(size = 10, sort = "id", direction = Sort.Direction.ASC) Pageable pageable) {
         Page<ProductResponse> productPage = productService.getAllProducts(pageable);
@@ -58,7 +44,7 @@ public class AdminController {
         return "products";
     }
 
-    @GetMapping("/products/new")
+    @GetMapping("/new")
     public String showAddProductForm(Model model) {
         List<CategoryResponse> categories = categoryService.getAllCategories();
         model.addAttribute("categories", categories);
@@ -66,7 +52,7 @@ public class AdminController {
         return "product_form";
     }
 
-    @PostMapping("/products")
+    @PostMapping("")
     public String addProduct(@Valid @ModelAttribute("product") ProductRequest productRequest,
         BindingResult result, Model model) {
         if (result.hasErrors()) {
@@ -78,7 +64,7 @@ public class AdminController {
         return "redirect:/admin/products";
     }
 
-    @GetMapping("/products/{id}/edit")
+    @GetMapping("/{id}/edit")
     public String showEditProductForm(@PathVariable("id") Long id, Model model) {
         ProductResponse productResponse = productService.getProductById(id);
         List<CategoryResponse> categories = categoryService.getAllCategories();
@@ -87,7 +73,7 @@ public class AdminController {
         return "product_edit";
     }
 
-    @PutMapping("/products/{id}")
+    @PutMapping("/{id}")
     public String updateProduct(@PathVariable("id") Long id,
         @Valid @ModelAttribute ProductRequest productRequest, BindingResult result, Model model) {
         if (result.hasErrors()) {
@@ -101,58 +87,9 @@ public class AdminController {
         return "redirect:/admin/products";
     }
 
-    @DeleteMapping("/products/{id}")
+    @DeleteMapping("/{id}")
     public String deleteProduct(@PathVariable("id") Long id) {
         productService.deleteProduct(id);
         return "redirect:/admin/products";
-    }
-
-
-    // 회원 관리
-    @GetMapping("/members")
-    public String getAllMembers(Model model) {
-        model.addAttribute("members", memberService.getAllMembers());
-        return "members";
-    }
-
-    @GetMapping("/members/new")
-    public String showAddMemberForm(Model model) {
-        model.addAttribute("member", new MemberRequest(null, "", ""));
-        return "member_form";
-    }
-
-    @PostMapping("/members")
-    public String addMember(@Valid @ModelAttribute("member") MemberRequest memberDTO,
-        BindingResult result) {
-        if (result.hasErrors()) {
-            return "member_form";
-        }
-        memberService.registerMember(memberDTO);
-        return "redirect:/admin/members";
-    }
-
-    @GetMapping("/members/{id}/edit")
-    public String showEditMemberForm(@PathVariable("id") Long id, Model model) {
-        MemberResponse memberResponse = memberService.getMemberById(id);
-        model.addAttribute("member", new Member(memberResponse.id(), memberResponse.email(), null));
-        return "member_edit";
-    }
-
-    @PutMapping("/members/{id}")
-    public String updateMember(@PathVariable("id") Long id,
-        @Valid @ModelAttribute MemberRequest memberRequest, BindingResult result, Model model) {
-        if (result.hasErrors()) {
-            model.addAttribute("member", memberRequest);
-            model.addAttribute("org.springframework.validation.BindingResult.member", result);
-            return "member_edit";
-        }
-        memberService.updateMember(id, memberRequest);
-        return "redirect:/admin/members";
-    }
-
-    @DeleteMapping("/members/{id}")
-    public String deleteMember(@PathVariable("id") Long id) {
-        memberService.deleteMember(id);
-        return "redirect:/admin/members";
     }
 }
